requirements.txt
boto3
amazon-textract-textractor
python-dotenv
requests
openai
groq
botocore
json5
pandas
langchain_groq
langchain
nltk